The start of moving the Mod to MC1.18.2 and Java 17
Many things are broken since the imports have changed to the official mapping. This branch may be on hold as I first update to 1.16.4 to just before the change to the official mappings. This will allow for a JDK update and MC version update before the whole library need a refactor. Also changed most of the dependencies to work with Gradle 7.2 and used Curse Maven to do most of the work of including the dependencies. This simplifies the need for including the other Maven repositories. Had to comment out processResources as it was not working correctly also I commented out uploadArchives as I do not own this maven repository.