The start of moving the Mod to MC1.15.2
Moving up the MC version to 1.15.2 and Forge 31.2.57. Mappings changed to stable_60-1.15 and this appears to have broken a few imports. Also changed most of the dependencies to work with Gradle 7.2 ForgeGradle 5.1.+ and used Curse Maven to do most of the work of including the dependencies. This simplifies the need for including the other Maven repositories. Had to comment out processResources as it was not working correctly also I commented out uploadArchives as I do not own this maven repository. As of this update there are 98 errors in the build. All this work should make updating to the next version of Minecraft easier.